[data] refactor: move sampler api to experimental (#2381)
