+++ tests/utils/dataset/test_create_rl_sampler_on_cpu.py
@@ -22,8 +22,8 @@ import torch
 from omegaconf import DictConfig, OmegaConf
 from torch.utils.data import Dataset, RandomSampler
 
+from verl.experimental.dataset.sampler import AbstractCurriculumSampler
 from verl.trainer.main_ppo import create_rl_sampler
-from verl.utils.dataset.sampler import AbstractCurriculumSampler
 
 
 class RandomCurriculumSampler(AbstractCurriculumSampler):
@@ -77,10 +77,11 @@ class MockChatDataset(Dataset):
 def test_create_custom_curriculum_samper():
     data_config = OmegaConf.create(
         {
+            "dataloader_num_workers": 0,
             "sampler": {
                 "class_path": "pkg://tests.utils.dataset.test_create_rl_sampler_on_cpu",
                 "class_name": "RandomCurriculumSampler",
-            }
+            },
         }
     )
